modules/whisper-service-mac/requirements.txt
# macOS-Optimized Whisper Service Requirements
# Native whisper.cpp integration with Apple Silicon optimizations

# Core Flask and web dependencies
Flask>=2.0.0
Flask-CORS>=3.0.0

# Audio processing (macOS optimized)
numpy>=1.21.0
soundfile>=0.10.0

# Configuration management
pyyaml>=6.0
python-dotenv>=0.19.0

# Performance monitoring
psutil>=5.8.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# macOS-specific audio support
# Note: Core ML and Metal are provided by macOS system frameworks

# whisper.cpp Python bindings (if using alternative to subprocess)
# whispercpp>=1.0.0  # Uncomment if using Python bindings instead of CLI

# Optional: Core ML model generation dependencies
# Uncomment these if you want to generate Core ML models:
# ane_transformers>=0.1.0
# openai-whisper>=20231117
# coremltools>=7.0

# Optional: Advanced audio processing
# librosa>=0.9.0
# pydub>=0.25.0

# Testing and development
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0